# schedularo-vibes-main/openapi.json
{"openapi":"3.1.0","info":{"title":"Appointment Booking System","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/verify-otp":{"post":{"tags":["Auth"],"summary":"Verify Otp","operationId":"verify_otp_auth_verify_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPVerifyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["Users"],"summary":"Get Current User Info","operationId":"get_current_user_info_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/appointment-types/":{"post":{"tags":["Appointment Types"],"summary":"Create","operationId":"create_appointment_types__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentTypeCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentTypeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/appointment-types/{appointment_id}/publish":{"post":{"tags":["Appointment Types"],"summary":"Publish","operationId":"publish_appointment_types__appointment_id__publish_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Appointment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/appointment-types/public":{"get":{"tags":["Appointment Types"],"summary":"List Published","operationId":"list_published_appointment_types_public_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AppointmentTypeResponse"},"type":"array","title":"Response List Published Appointment Types Public Get"}}}}}}},"/working-hours/":{"post":{"tags":["Working Hours"],"summary":"Add Working Hours","operationId":"add_working_hours_working_hours__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkingHoursCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/slots/generate":{"post":{"tags":["Slots"],"summary":"Generate Slots Api","operationId":"generate_slots_api_slots_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SlotGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/slots":{"get":{"tags":["Slots"],"summary":"Get Slots","operationId":"get_slots_slots_get","parameters":[{"name":"appointment_type_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","description":"Appointment type ID","title":"Appointment Type Id"},"description":"Appointment type ID"},{"name":"slot_date","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date to get slots for","title":"Slot Date"},"description":"Date to get slots for"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SlotResponse"},"title":"Response Get Slots Slots Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resources/":{"post":{"tags":["Resources"],"summary":"Create","operationId":"create_resources__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/resources/{resource_id}/toggle":{"post":{"tags":["Resources"],"summary":"Toggle","operationId":"toggle_resources__resource_id__toggle_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"resource_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Resource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resources/assign":{"post":{"tags":["Resources"],"summary":"Assign","operationId":"assign_resources_assign_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"appointment_type_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Appointment Type Id"}},{"name":"resource_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Resource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resources/appointment/{appointment_type_id}":{"get":{"tags":["Resources"],"summary":"List For Appointment","operationId":"list_for_appointment_resources_appointment__appointment_type_id__get","parameters":[{"name":"appointment_type_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Appointment Type Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResourceResponse"},"title":"Response List For Appointment Resources Appointment  Appointment Type Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bookings/":{"post":{"tags":["Bookings"],"summary":"Book Slot","operationId":"book_slot_bookings__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/bookings/my":{"get":{"tags":["Bookings"],"summary":"Get My Bookings","operationId":"get_my_bookings_bookings_my_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookingResponse"},"type":"array","title":"Response Get My Bookings Bookings My Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/bookings/all":{"get":{"tags":["Bookings"],"summary":"Get All Bookings","operationId":"get_all_bookings_bookings_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookingResponse"},"type":"array","title":"Response Get All Bookings Bookings All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/payments/":{"post":{"tags":["Payments"],"summary":"Pay","operationId":"pay_payments__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/appointments/{appointment_id}/cancel":{"post":{"tags":["Cancellation"],"summary":"Cancel","operationId":"cancel_appointments__appointment_id__cancel_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Appointment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/appointment-questions/":{"post":{"tags":["Appointment Questions"],"summary":"Add Question","operationId":"add_question_appointment_questions__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentQuestionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/appointment-questions/{appointment_type_id}":{"get":{"tags":["Appointment Questions"],"summary":"Get Questions","operationId":"get_questions_appointment_questions__appointment_type_id__get","parameters":[{"name":"appointment_type_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Appointment Type Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppointmentQuestionResponse"},"title":"Response Get Questions Appointment Questions  Appointment Type Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/appointment-answers/":{"post":{"tags":["Appointment Answers"],"summary":"Submit","operationId":"submit_appointment_answers__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentAnswerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/appointment-status/{appointment_id}":{"get":{"tags":["Appointment Audit"],"summary":"Get History","operationId":"get_history_appointment_status__appointment_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"title":"Appointment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AppointmentAnswerCreate":{"properties":{"appointment_id":{"type":"string","format":"uuid","title":"Appointment Id"},"question_id":{"type":"string","format":"uuid","title":"Question Id"},"answer":{"type":"string","title":"Answer"}},"type":"object","required":["appointment_id","question_id","answer"],"title":"AppointmentAnswerCreate"},"AppointmentQuestionCreate":{"properties":{"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"question_text":{"type":"string","title":"Question Text"},"input_type":{"type":"string","title":"Input Type","default":"text"},"is_required":{"type":"boolean","title":"Is Required","default":false}},"type":"object","required":["appointment_type_id","question_text"],"title":"AppointmentQuestionCreate"},"AppointmentQuestionResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"question_text":{"type":"string","title":"Question Text"},"input_type":{"type":"string","title":"Input Type"},"is_required":{"type":"boolean","title":"Is Required"}},"type":"object","required":["id","appointment_type_id","question_text","input_type","is_required"],"title":"AppointmentQuestionResponse"},"AppointmentTypeCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"duration_minutes":{"type":"integer","title":"Duration Minutes"},"appointment_mode":{"type":"string","title":"Appointment Mode"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"}},"type":"object","required":["name","description","duration_minutes","appointment_mode","location"],"title":"AppointmentTypeCreate"},"AppointmentTypeResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"duration_minutes":{"type":"integer","title":"Duration Minutes"},"appointment_mode":{"type":"string","title":"Appointment Mode"},"is_published":{"type":"boolean","title":"Is Published"}},"type":"object","required":["id","name","duration_minutes","appointment_mode","is_published"],"title":"AppointmentTypeResponse"},"BookingCreate":{"properties":{"slot_id":{"type":"string","format":"uuid","title":"Slot Id"},"people_count":{"type":"integer","title":"People Count","default":1}},"type":"object","required":["slot_id"],"title":"BookingCreate"},"BookingResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"slot_id":{"type":"string","format":"uuid","title":"Slot Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"people_count":{"type":"integer","title":"People Count"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"appointment_type_name":{"type":"string","title":"Appointment Type Name","default":""},"slot_date":{"type":"string","title":"Slot Date","default":""},"start_time":{"type":"string","title":"Start Time","default":""},"end_time":{"type":"string","title":"End Time","default":""},"booked_by":{"type":"string","title":"Booked By","default":""}},"type":"object","required":["id","appointment_type_id","slot_id","user_id","people_count","status","created_at"],"title":"BookingResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"OTPVerifyRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["email","otp"],"title":"OTPVerifyRequest"},"PaymentCreate":{"properties":{"appointment_id":{"type":"string","format":"uuid","title":"Appointment Id"},"amount":{"type":"number","title":"Amount"},"payment_method":{"type":"string","title":"Payment Method"}},"type":"object","required":["appointment_id","amount","payment_method"],"title":"PaymentCreate"},"RegisterRequest":{"properties":{"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"CUSTOMER"}},"type":"object","required":["full_name","email","password"],"title":"RegisterRequest"},"ResourceCreate":{"properties":{"name":{"type":"string","title":"Name"},"capacity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Capacity","default":1}},"type":"object","required":["name"],"title":"ResourceCreate"},"ResourceResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"capacity":{"type":"integer","title":"Capacity"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","name","capacity","is_active"],"title":"ResourceResponse"},"SlotGenerateRequest":{"properties":{"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"slot_date":{"type":"string","format":"date","title":"Slot Date"}},"type":"object","required":["appointment_type_id","slot_date"],"title":"SlotGenerateRequest"},"SlotResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"slot_date":{"type":"string","format":"date","title":"Slot Date"},"start_time":{"type":"string","format":"time","title":"Start Time"},"end_time":{"type":"string","format":"time","title":"End Time"},"max_capacity":{"type":"integer","title":"Max Capacity"},"booked_capacity":{"type":"integer","title":"Booked Capacity"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","appointment_type_id","slot_date","start_time","end_time","max_capacity","booked_capacity","status"],"title":"SlotResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"},"is_verified":{"type":"boolean","title":"Is Verified"},"roles":{"items":{"type":"string"},"type":"array","title":"Roles","default":[]}},"type":"object","required":["id","full_name","email","is_active","is_verified"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkingHoursCreate":{"properties":{"appointment_type_id":{"type":"string","format":"uuid","title":"Appointment Type Id"},"weekday":{"type":"integer","title":"Weekday"},"start_time":{"type":"string","format":"time","title":"Start Time"},"end_time":{"type":"string","format":"time","title":"End Time"}},"type":"object","required":["appointment_type_id","weekday","start_time","end_time"],"title":"WorkingHoursCreate"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}